Add initial docs for Template Library
In this commit we are paving the way for establishing the very first
documentation for Template Library. The documentation is currently in
the form of .rst files and is prepared to be build with Read the
Docs and Sphinx documentation tool generator. With these changes
we set the structure for hosting our docs on GitHub Pages.